Store and retrieve block type impressions from AppPrefs
We track changes to the block type impressions to appropriate remove the
"new" badge for a give block type.

+++ libs/editor/WordPressEditor/src/main/java/org/wordpress/android/editor/EditorFragmentAbstract.java
@@ -222,8 +222,8 @@ public abstract class EditorFragmentAbstract extends Fragment {
         void onReplaceStoryEditedBlockActionSent();
         void onReplaceStoryEditedBlockActionReceived();
         boolean showPreview();
-        Map<String, Integer> onRequestBlockTypeImpressions(ArrayList<Object> newBlockTypes);
-        void onSetBlockTypeImpressionCount(String name, Integer count);
+        Map<String, Double> onRequestBlockTypeImpressions(ArrayList<Object> newBlockTypes);
+        void onSetBlockTypeImpressionCount(String name, Double count);
     }
 
     /**
